Add win conditions and logic to input players tokens.
Created a function that returns true if the one of eight winning
conditions (as represented by player tokens taking up parts of the
board array). I feel like there must be a more elegant solution than
manually listing all eight conditions, but I'll go with it for now.

The logic to input player tokens uses the getActivePlayer method,
which checks if the player object has the boolean property of
activePlayer. The activePlayer's playerToken property is then passed
to the setPosition method.