Fix bug preventing winning conditions from executing.
The issue here was that isXwinner and isOWinner were being initialized
(and initialized as false), but not called again in the course of the
game. This was fixed by moving the winningConditions() call to the
handleClick event handler (though these properties were initialized)
outside of the event handler in order to make them avaiable to the
return statement.